+++ src/Configuration/Dotenv/Scheme/Parser.hs
@@ -1,7 +1,24 @@
-{-# LANGUAGE CPP               #-}
-{-# LANGUAGE OverloadedStrings #-}
+-- |
+-- Module      :  Configuration.Dotenv.Types
+-- Copyright   :  © 2015–2018 Stack Builders Inc.
+-- License     :  MIT
+--
+-- Maintainer  :  Stack Builders <hackage@stackbuilders.com>
+-- Stability   :  experimental
+-- Portability :  portable
+--
+-- Helpers for loadSafeFile
 
-module Configuration.Dotenv.Scheme.Parser where
+{-# LANGUAGE CPP                 #-}
+{-# LANGUAGE OverloadedStrings   #-}
+{-# LANGUAGE ScopedTypeVariables #-}
+
+module Configuration.Dotenv.Scheme.Parser
+  ( defaultValidatorMap
+  , parseEnvsWithScheme
+  , typeValidator
+  )
+  where
 
 #if !MIN_VERSION_base(4,8,0)
 import Control.Applicative ((*>), pure)
@@ -9,39 +26,40 @@ import Data.Functor ((<$>))
 #endif
 
 import Data.Either
-import Data.Void
-import Text.Megaparsec
-import Text.Megaparsec.Char
+
+import qualified Data.Map.Lazy as ML
+
+import qualified Data.Text as T
 
 import Configuration.Dotenv.Scheme.Types
 
+
+-- |
+--
 parseEnvsWithScheme
-  :: [(String, EnvType)]
-  -> Either [ParseError Char Void] ()
-parseEnvsWithScheme valuesAndTypes =
+  :: ValidatorMap -- ^ MapFormat validations
+  -> [(String, EnvType)] -- ^ Value and Type
+  -> Either [String] ()
+parseEnvsWithScheme validatorMap valuesAndTypes =
   let parsedEithers = parseEnvs <$> valuesAndTypes
-      parseEnvs (val, type') = val `parseEnvAs` type'
+      parseEnvWith = typeValidator validatorMap
+      parseEnvs (val, type') = val `parseEnvWith` type'
     in if all isRight parsedEithers
           then Right ()
           else Left (lefts parsedEithers)
 
-parseEnvAs
-  :: String -- ^ Value of the env variable
+-- |
+--
+typeValidator
+  :: ValidatorMap -- ^ MapFormat validations
+  -> String -- ^ Value of the env variable
   -> EnvType -- ^ Type that the env variable should have
-  -> Either (ParseError Char Void) ()
-parseEnvAs envVal envTypeNeeded =
-  parse dispatch "" envVal
-    where
-      errorMsg = "Couldn't parse " ++ envVal ++ " as " ++ showType envTypeNeeded
-      evalParse parser = parser *> eof *> pure () <|> fail errorMsg
-      dispatch :: Parsec Void String ()
-      dispatch =
-        case envTypeNeeded of
-          EnvInteger -> evalParse (many digitChar)
-          EnvBool    -> evalParse (string "true" <|> string "false")
-          EnvText    -> pure ()
-
-showType :: EnvType -> String
-showType EnvInteger = "integer"
-showType EnvBool = "boolean"
-showType _ = "text"
+  -> Either String ()
+typeValidator validatorMap envVal (EnvType type_) =
+  let errorMsg = "Couldn't parse " ++ envVal ++ " as " ++ T.unpack type_
+   in case ML.lookup type_ validatorMap of
+        Nothing -> Left errorMsg
+        Just validator ->
+          if validator (T.pack envVal)
+             then Right ()
+             else Left errorMsg