+++ src/Configuration/Dotenv/Scheme/Types.hs
@@ -1,3 +1,14 @@
+-- |
+-- Module      :  Configuration.Dotenv.Types
+-- Copyright   :  © 2015–2018 Stack Builders Inc.
+-- License     :  MIT
+--
+-- Maintainer  :  Stack Builders <hackage@stackbuilders.com>
+-- Stability   :  experimental
+-- Portability :  portable
+--
+-- Types for 'loadSafeFile' (e. g., 'ValidatorMap')
+
 {-# LANGUAGE CPP #-}
 {-# LANGUAGE OverloadedStrings #-}
 
@@ -8,21 +19,33 @@ import Control.Applicative ((<*>), pure)
 import Data.Functor ((<$>))
 #endif
 
+import Data.Maybe (isJust)
+
+import Data.Map.Lazy (Map)
+
 import Data.Yaml
 
-data EnvType =
-  EnvInteger
-    | EnvBool
-    | EnvText
+import qualified Data.Map.Lazy as ML
+
+import Data.Text (Text)
+import qualified Data.Text as T
+
+import Text.Read (readMaybe)
+
+
+-- |
+--
+newtype EnvType = EnvType Text
     deriving (Show, Eq, Ord)
 
+-- |
+--
 instance FromJSON EnvType where
-  parseJSON (String "integer") = pure EnvInteger
-  parseJSON (String "bool") = pure EnvBool
-  parseJSON (String "text") = pure EnvText
-  parseJSON (String x) = fail ("Don't know how to parse that kind of type: " ++ show x)
-  parseJSON x = fail ("Not an object: " ++ show x)
+  parseJSON (String value) = pure (EnvType value)
+  parseJSON anyOther = fail ("Not an object: " ++ show anyOther)
 
+-- |
+--
 data Env =
   Env
     { envName  :: String
@@ -30,6 +53,8 @@ data Env =
     , required :: Bool
     } deriving (Show, Eq, Ord)
 
+-- |
+--
 instance FromJSON Env where
   parseJSON (Object m) =
     Env
@@ -37,3 +62,29 @@ instance FromJSON Env where
       <*> m .: "type"
       <*> m .:? "required" .!= False
   parseJSON x = fail ("Not an object: " ++ show x)
+
+-- | Parameters:
+--
+-- - __Key:__ Name of the /format/ to check.
+--
+-- - __Value:__ Function to check if some text meets the condition.
+--
+type ValidatorMap = Map Text (Text -> Bool)
+
+
+-- | Default configuration for 'loadSafeFile'. It currently checks:
+-- @bool@, @integer@, and @text@.
+--
+defaultValidatorMap :: ValidatorMap
+defaultValidatorMap =
+  let booleanValidator :: Text -> Bool
+      booleanValidator text = isJust (readMaybe (T.unpack text) :: Maybe Bool)
+      integerValidator :: Text -> Bool
+      integerValidator text = isJust (readMaybe (T.unpack text) :: Maybe Integer)
+      textValidator :: Text -> Bool
+      textValidator = const True
+   in ML.fromList
+        [ ("bool", booleanValidator)
+        , ("integer", integerValidator)
+        , ("text", textValidator)
+        ]